Provide access to case status (S_STATUS) field
The FB API will only return fields that are explicitly requested.

Our usage is interested in the status of each case, which is provided
by FB in the sStatus field.

This change adds a constant S_STATUS for this field, and requests it
when querying a case.

The constructors for Case are updated to allow the extra status string
argument. This doesn't feel like the best way to extend the class.

A getStatus() method is added to return the status string. No
setStatus() is provided, as we don't (currently) need that.

Signed-off-by: Chris Ridd <chrisridd@mac.com>